% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCMD.R
\name{simCMD}
\alias{simCMD}
\title{simulation d'une CM à temps discret}
\usage{
simCMD(x, ei, Pt, n)
}
\arguments{
\item{x}{numerical vector}

\item{ei}{numerical vector}

\item{Pt}{numerical matrix}

\item{n}{integer}
}
\description{
simulation d'une CM à temps discret
}
